split chatbox out of main game view
